+++ examples/resources/konnect_cloud_gateway_add_on/import.sh
@@ -1,1 +1,0 @@
-terraform import konnect_cloud_gateway_add_on.my_konnect_cloud_gateway_add_on "550e8400-e29b-41d4-a716-446655440000"