+++ examples/resources/konnect_cloud_gateway_addon/import-by-string-id.tf
@@ -1,5 +1,5 @@
 import {
   provider = konnect-beta
-  to = konnect_cloud_gateway_add_on.my_konnect-beta_cloud_gateway_add_on
+  to = konnect_cloud_gateway_addon.my_konnect-beta_cloud_gateway_addon
   id = "550e8400-e29b-41d4-a716-446655440000"
 }