+++ internal/provider/provider.go
@@ -163,7 +163,7 @@ func (p *KonnectBetaProvider) Resources(ctx context.Context) []func() resource.R
 		NewAuthServerClaimsResource,
 		NewAuthServerClientsResource,
 		NewAuthServerScopesResource,
-		NewCloudGatewayAddOnResource,
+		NewCloudGatewayAddonResource,
 		NewDashboardResource,
 		NewEventGatewayResource,
 		NewEventGatewayBackendClusterResource,